CRAG/pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Sudeshna Pathak",email = "pathaksudeshna92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.3,<0.5.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "tavily-python (>=0.7.2,<0.8.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "langchain-chroma (>=0.2.3,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
